Move out LBSyncRefreshEventData to proper file
