+++ common/synchronization/synchronization-core/src/commonMain/kotlin/studio/lunabee/synchronization/syncmanager/LBSyncRefreshEventData.kt
@@ -1,0 +1,31 @@
+/*
+ * Copyright (c) 2026 Lunabee Studio
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package studio.lunabee.synchronization.syncmanager
+
+import kotlin.reflect.KClass
+
+sealed interface LBSyncRefreshEventData {
+    val type: KClass<out LBSyncRefreshEvent>
+
+    data class AppForeground(val isForeground: Boolean) : LBSyncRefreshEventData {
+        override val type: KClass<out LBSyncRefreshEvent> = LBSyncRefreshEvent.AppForeground::class
+    }
+
+    data object InternetIsBack : LBSyncRefreshEventData {
+        override val type: KClass<out LBSyncRefreshEvent> = LBSyncRefreshEvent.InternetIsBack::class
+    }
+}
